Add files via upload
Web page designed for "Responsive Web Design Projects - Technical Documentation Page" on FreeCodeCamp.org. See the link below for a description of the user stories that were targeted within the project. At the time of its creation, I was not versed well enough in any one thing to write a technical document, so I created a very simplified curriculum for an English as a second language program.

https://learn.freecodecamp.org/responsive-web-design/responsive-web-design-projects/build-a-technical-documentation-page/